+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 
 ## 0.2.1
 * adds support for cloth-config2, modmenu
-* adds options screen/json
+* add options screen/json
 * adds option to disable backpack trinket even when installed
+* change how trinkets are updated
 
 ## 0.2.0
 * adds support for trinkets